requirements.txt
# MLOps Core
mlflow==2.10.2
feast[redis,duckdb]==0.40.0
bentoml==1.2.3
apache-airflow==2.10.2

# Data Processing
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
xgboost==2.0.3
duckdb>=0.10.0
pyarrow==15.0.0
shap==0.43.0

# Monitoring
evidently>=0.5.0
prometheus-client==0.19.0

# GCP
google-cloud-bigquery==3.15.0
google-cloud-storage==2.14.0
google-cloud-secret-manager==2.17.0

# Streaming
confluent-kafka>=2.3.0
faust-streaming>=0.11.0

# Utilities
pydantic==2.5.3
python-dotenv==1.0.0
redis>=4.2.2,<5
psycopg2-binary==2.9.9
boto3==1.34.34
optuna==3.6.1
python-multipart>=0.0.6

